requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
aiohttp==3.10.0
beautifulsoup4==4.12.3
aiosqlite==0.20.0
apscheduler==3.10.4
jinja2==3.1.4
python-multipart==0.0.9
pyjwt==2.9.0
bcrypt==4.2.0
